+++ database/seeders/DatabaseSeeder.php
@@ -17,8 +17,10 @@ class DatabaseSeeder extends Seeder
         $this->call(DeveloperAccessSeeder::class);
         $this->call(SettingGroupTableSeeder::class);
         $this->call(SettingTableSeeder::class);
-        $this->call(LocationSeeder::class);        
-        $this->call(DeveloperAccessSeeder::class);
         $this->call(RolesandPermissionsSeeder::class);
+        $this->call(LocationSeeder::class);
+        $this->call(FaqCategoriesSeeder::class);
+        $this->call(FaqEntriesSeeder::class);
+        $this->call(CausesSeeder::class);
     }
 }